# hms-project/docs/ambulance-management-openapi.yaml
openapi: 3.0.3
info:
  title: HMS Ambulance Management API
  description: >-
    This OpenAPI specification documents all ambulance management related
    endpoints, including ambulance inventory management, dispatch system,
    real-time location tracking, driver management, service history,
    and billing integration.
  version: 1.0.0
servers:
  - url: https://hms.example.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local dev server
...
paths:
...
  /ambulances/billing:
    get:
      tags: [Ambulance Billing]
      summary: Get ambulance billing records with optional filtering
      parameters:
        - in: query
          name: id
          schema:
            type: string
        - in: query
          name: dispatchId
          schema:
            type: string
        - in: query
          name: paymentStatus
          schema:
            type: string
            enum: [PENDING, PAID, PARTIAL, CANCELLED, INSURANCE_PENDING]
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: List of ambulance billing records with pagination metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingListResponse'
    post:
      tags: [Ambulance Billing]
      summary: Create a new ambulance billing record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillingInput'
      responses:
        '201':
          description: Billing record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Billing'
    put:
      tags: [Ambulance Billing]
      summary: Update an ambulance billing record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/BillingInput'
                - type: object
                  properties:
                    id:
                      type: string
                  required: [id]
      responses:
        '200':
          description: Billing record updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Billing'
  /ambulances/billing/calculate:
    post:
      tags: [Ambulance Billing]
      summary: Calculate ambulance billing based on various parameters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillingCalculationInput'
      responses:
        '200':
          description: Calculated billing breakdown
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingCalculationResult'
components:
  schemas:
    PagedResponse:
      type: object
      properties:
        data:
          type: array
          items: { type: object }
        meta:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    BillingListResponse:
      allOf:
        - $ref: '#/components/schemas/PagedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Billing'
    BillingInput:
      type: object
      required: [dispatchId, baseCharge, distanceCharge, totalAmount]
      properties:
        dispatchId:
          type: string
        baseCharge:
          type: number
          format: double
        distanceCharge:
          type: number
          format: double
        waitingCharge:
          type: number
          format: double
        equipmentCharge:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        insuranceCovered:
          type: boolean
          default: false
        insuranceProvider:
          type: string
        insurancePolicyNumber:
          type: string
        notes:
          type: string
    Billing:
      allOf:
        - $ref: '#/components/schemas/BillingInput'
        - type: object
          properties:
            id:
              type: string
            paymentStatus:
              type: string
              enum: [PENDING, PAID, PARTIAL, CANCELLED, INSURANCE_PENDING]
            invoiceNumber:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    BillingCalculationInput:
      type: object
      required: [distance, vehicleType, serviceLevel]
      properties:
        distance:
          type: number
          format: double
        vehicleType:
          type: string
          enum: [BASIC_LIFE_SUPPORT, ADVANCED_LIFE_SUPPORT, PATIENT_TRANSPORT, NEONATAL, MOBILE_ICU]
        serviceLevel:
          type: string
          enum: [BASIC, STANDARD, PREMIUM]
        timeOfDay:
          type: string
          enum: [NORMAL, NIGHT, HOLIDAY]
          default: NORMAL
        waitingTime:
          type: integer
          default: 0
        zone:
          type: string
          enum: [URBAN, SUBURBAN, RURAL, HIGHWAY]
          default: URBAN
        equipment:
          type: array
          items:
            type: string
        medicalStaff:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              count:
                type: integer
        insuranceInfo:
          type: object
          properties:
            provider:
              type: string
            policyNumber:
              type: string
            coverageType:
              type: string
              enum: [BASIC, STANDARD, PREMIUM, FULL]
    BillingCalculationResult:
      type: object
      properties:
        breakdown:
          type: object
          properties:
            baseRate:
              type: number
              format: double
            distanceCharge:
              type: number
              format: double
            waitingCharge:
              type: number
              format: double
            equipmentCharge:
              type: number
              format: double
            staffCharge:
              type: number
              format: double
            subtotal:
              type: number
              format: double
            insuranceCoverage:
              type: number
              format: double
            patientResponsibility:
              type: number
              format: double
            tax:
              type: number
              format: double
            totalAmount:
              type: number
              format: double
        rates:
          type: object
          properties:
            baseRate:
              type: number
              format: double
            distanceRate:
              type: number
              format: double
            timeAdjustment:
              type: number
              format: double
            gstRate:
              type: number
              format: double